requirements.txt
streamlit==1.37.0
Pillow==10.4.0
numpy==1.26.4
